# student/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Student API
    version: 0.0.1
paths:
    /student:
        post:
            tags:
                - Student
            operationId: Student_CreateStudent
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateStudentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateStudentReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /student/{id}:
        get:
            tags:
                - Student
            operationId: Student_GetStudent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetStudentReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        put:
            tags:
                - Student
            operationId: Student_UpdateStudent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateStudentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateStudentReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - Student
            operationId: Student_DeleteStudent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteStudentReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /students:
        get:
            tags:
                - Student
            operationId: Student_ListStudent
            parameters:
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListStudentReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        CreateStudentReply:
            type: object
            properties:
                name:
                    type: string
                status:
                    type: integer
                    format: int32
                id:
                    type: integer
                    format: int32
                info:
                    type: string
        CreateStudentRequest:
            type: object
            properties:
                name:
                    type: string
                status:
                    type: integer
                    format: int32
                info:
                    type: string
        DeleteStudentReply:
            type: object
            properties: {}
        GetStudentReply:
            type: object
            properties:
                name:
                    type: string
                status:
                    type: integer
                    format: int32
                id:
                    type: integer
                    format: int32
                info:
                    type: string
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        ListStudentReply:
            type: object
            properties:
                results:
                    type: array
                    items:
                        $ref: '#/components/schemas/ListStudentReply_Result'
        ListStudentReply_Result:
            type: object
            properties:
                name:
                    type: string
                status:
                    type: integer
                    format: int32
                id:
                    type: integer
                    format: int32
                info:
                    type: string
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UpdateStudentReply:
            type: object
            properties:
                name:
                    type: string
                status:
                    type: integer
                    format: int32
                id:
                    type: integer
                    format: int32
                info:
                    type: string
        UpdateStudentRequest:
            type: object
            properties:
                name:
                    type: string
                status:
                    type: integer
                    format: int32
                id:
                    type: integer
                    format: int32
                info:
                    type: string
tags:
    - name: Student